split HistoryWork files into per-class files
