+++ src/history/HistoryArchive.cpp
@@ -9,9 +9,9 @@
 #include "history/HistoryArchive.h"
 #include "StellarCoreVersion.h"
 #include "bucket/BucketList.h"
+#include "bucket/Bucket.h"
 #include "crypto/Hex.h"
 #include "crypto/SHA.h"
-#include "history/FileTransferInfo.h"
 #include "history/HistoryManager.h"
 #include "lib/util/format.h"
 #include "main/Application.h"
