+++ src/history/HistoryManagerImpl.cpp
@@ -12,11 +12,17 @@
 #include "crypto/Hex.h"
 #include "crypto/SHA.h"
 #include "herder/HerderImpl.h"
-#include "history/FileTransferInfo.h"
 #include "history/HistoryArchive.h"
 #include "history/HistoryManagerImpl.h"
-#include "history/HistoryWork.h"
 #include "history/StateSnapshot.h"
+#include "historywork/CatchupCompleteWork.h"
+#include "historywork/CatchupMinimalWork.h"
+#include "historywork/CatchupRecentWork.h"
+#include "historywork/FetchRecentQsetsWork.h"
+#include "historywork/GetHistoryArchiveStateWork.h"
+#include "historywork/PublishWork.h"
+#include "historywork/PutHistoryArchiveStateWork.h"
+#include "historywork/RepairMissingBucketsWork.h"
 #include "ledger/LedgerManager.h"
 #include "lib/util/format.h"
 #include "main/Application.h"
@@ -30,6 +36,7 @@
 #include "util/StatusManager.h"
 #include "util/TmpDir.h"
 #include "util/make_unique.h"
+#include "work/WorkManager.h"
 #include "xdrpp/marshal.h"
 
 #include <fstream>
