+++ src/history/HistoryTests.cpp
@@ -9,7 +9,10 @@
 #include "herder/LedgerCloseData.h"
 #include "history/HistoryArchive.h"
 #include "history/HistoryManager.h"
-#include "history/HistoryWork.h"
+#include "historywork/GetHistoryArchiveStateWork.h"
+#include "historywork/GunzipFileWork.h"
+#include "historywork/GzipFileWork.h"
+#include "historywork/PutHistoryArchiveStateWork.h"
 #include "ledger/LedgerManager.h"
 #include "lib/catch.hpp"
 #include "main/Application.h"
